requirements.txt
Django==5.2
python-dotenv==1.0.0
djangorestframework==3.15.0
django-cors-headers==4.3.1 